+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file. This projec
 
 [Full Changelog](https://github.com/Shatur95/optimus-manager-qt/compare/1.2.2...HEAD)
 
-_No unreleased changes yet_
+**Changed**
+
+-   Fix memory leak in DaemonClient.
 
 ## [1.2.2](https://github.com/Shatur95/optimus-manager-qt/tree/1.2.2) (2019-06-04)